pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alpha-ai"
version = "0.1.0"
description = "Agentic AI offensive-security framework: MCP server + REST API exposing security tools to LLMs"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "Krishcalin" }]
keywords = ["security", "pentesting", "mcp", "ai", "red-team"]

dependencies = [
    "fastmcp>=0.2.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
    "structlog>=24.0.0",
    "rich>=13.7.0",
    "pyyaml>=6.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "ruff>=0.4.0",
    "mypy>=1.9",
]

[project.scripts]
alpha-mcp = "alpha_ai.servers.mcp_server:main"
alpha-api = "alpha_ai.servers.rest_api:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["alpha_ai*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py311"
